test(frontend): reusable mock Data
Refactor mock API data into reusable mock objects, request handlers and custom hooks.
