+++ frontend-v2/src/stories/Model.stories.tsx
@@ -4,34 +4,21 @@ import { expect, screen, within, fn, waitFor } from "storybook/test";
 import { useDispatch } from "react-redux";
 import { setProject as setReduxProject } from "../features/main/mainSlice";
 
-import {
-  CombinedModelUpdate,
-  ProjectUpdate,
-  TabbedModelForm,
-} from "../features/model/Model";
+import { TabbedModelForm } from "../features/model/Model";
 import {
   model,
-  pd_model,
   project,
   simulation,
-  variables,
   protocols,
   compound,
+  variables,
   units,
-  pkModels,
-  pdModels,
-} from "./model.mock";
-import {
-  DerivedVariableRead,
-  TimeIntervalRead,
-  useCombinedModelUpdateMutation,
-  useProjectUpdateMutation,
-} from "../app/backendApi";
-import { useEffect, useState } from "react";
-
-// Use mutable copies of snapshots to allow for API updates.
-let mockModel = { ...model };
-let mockProject = { ...project };
+  projectHandlers,
+  useMockModel,
+  useMockProject,
+} from "./project.mock";
+import { pd_model, pkModels, pdModels } from "./model.mock";
+import { useEffect } from "react";
 
 const meta: Meta<typeof TabbedModelForm> = {
   title: "Edit Model",
@@ -52,6 +39,7 @@ const meta: Meta<typeof TabbedModelForm> = {
     layout: "fullscreen",
     msw: {
       handlers: [
+        ...projectHandlers,
         http.get("/api/pharmacokinetic", async () => {
           await delay();
           return HttpResponse.json(pkModels, {
@@ -61,101 +49,6 @@ const meta: Meta<typeof TabbedModelForm> = {
         http.get("/api/pharmacodynamic", async () => {
           await delay();
           return HttpResponse.json(pdModels, {
-            status: 200,
-          });
-        }),
-        http.get("/api/combined_model", async ({ request }) => {
-          await delay();
-          const url = new URL(request.url);
-
-          const projectId = url.searchParams.get("project_id");
-          if (projectId) {
-            return HttpResponse.json([mockModel], {
-              status: 200,
-            });
-          }
-          return HttpResponse.json([], { status: 200 });
-        }),
-        http.get("/api/unit", async ({ request }) => {
-          await delay();
-          const url = new URL(request.url);
-
-          const compoundId = url.searchParams.get("compound_id");
-          if (compoundId) {
-            return HttpResponse.json(units, {
-              status: 200,
-            });
-          }
-          return HttpResponse.json([], { status: 200 });
-        }),
-        http.get("/api/variable", async ({ request }) => {
-          await delay();
-          const url = new URL(request.url);
-          const pkModel = url.searchParams.get("dosed_pk_model_id");
-          if (pkModel) {
-            return HttpResponse.json(variables, {
-              status: 200,
-            });
-          }
-          return HttpResponse.json([], { status: 200 });
-        }),
-        http.get("/api/project/:id", async ({ params }) => {
-          await delay();
-          //@ts-expect-error params.id is a string
-          const projectId = parseInt(params.id, 10);
-          if (projectId === mockProject.id) {
-            return HttpResponse.json(mockProject, {
-              status: 200,
-            });
-          }
-          return HttpResponse.json(
-            { error: "Project not found" },
-            { status: 404 },
-          );
-        }),
-        http.put("/api/combined_model/:id", async ({ params, request }) => {
-          await delay();
-          //@ts-expect-error params.id is a string
-          const modelId = parseInt(params.id, 10);
-          const modelData = await request.json();
-          //@ts-expect-error modelData is DefaultBodyType
-          const timeIntervals = modelData?.time_intervals.map(
-            (interval: TimeIntervalRead, index: number) => {
-              return {
-                ...interval,
-                pkpd_model: modelId,
-                id: interval.id || index + 1, // Ensure each interval has a unique ID
-              };
-            },
-          );
-          //@ts-expect-error modelData is DefaultBodyType
-          const derivedVariables = modelData?.derived_variables.map(
-            (variable: DerivedVariableRead, index: number) => {
-              return {
-                ...variable,
-                id: variable.id || index + 1, // Ensure each derived variable has a unique ID
-              };
-            },
-          );
-          mockModel = {
-            //@ts-expect-error modelData is DefaultBodyType
-            ...modelData,
-            id: modelId,
-            time_intervals: timeIntervals,
-            derived_variables: derivedVariables,
-          };
-          return HttpResponse.json(mockModel, {
-            status: 200,
-          });
-        }),
-        http.put("/api/project/:id", async ({ params, request }) => {
-          await delay();
-          //@ts-expect-error params.id is a string
-          const projectId = parseInt(params.id, 10);
-          const projectData = await request.json();
-          //@ts-expect-error projectData is DefaultBodyType
-          mockProject = { ...projectData, id: projectId };
-          return HttpResponse.json(mockProject, {
             status: 200,
           });
         }),
@@ -172,29 +65,13 @@ const meta: Meta<typeof TabbedModelForm> = {
             },
           );
         }),
-        http.put(
-          "/api/combined_model/:id/set_params_to_defaults",
-          async ({ params, request }) => {
-            await delay();
-            //@ts-expect-error params.id is a string
-            const modelId = parseInt(params.id, 10);
-            const modelData = await request.json();
-            //@ts-expect-error modelData is DefaultBodyType
-            mockModel = { ...modelData, id: modelId };
-            return HttpResponse.json(mockModel, {
-              status: 200,
-            });
-          },
-        ),
       ],
     },
   },
   decorators: [
     (Story, { args }) => {
-      const [model, setModel] = useState(args.model);
-      const [project, setProject] = useState(args.project);
-      const [updateModel] = useCombinedModelUpdateMutation();
-      const [updateProject] = useProjectUpdateMutation();
+      const [model, updateMockModel] = useMockModel(args);
+      const [project, updateMockProject] = useMockProject(args);
       const dispatch = useDispatch();
       const projectId = args.project.id;
 
@@ -202,23 +79,6 @@ const meta: Meta<typeof TabbedModelForm> = {
         dispatch(setReduxProject(projectId));
       }, [dispatch, projectId]);
 
-      const updateMockModel: CombinedModelUpdate = async ({
-        id,
-        combinedModel,
-      }) => {
-        args.updateModel({ id, combinedModel });
-        await updateModel({ id, combinedModel });
-        mockModel = { ...mockModel, ...combinedModel };
-        setModel(mockModel);
-        return mockModel;
-      };
-      const updateMockProject: ProjectUpdate = async ({ id, project }) => {
-        args.updateProject({ id, project });
-        await updateProject({ id, project });
-        mockProject = { ...mockProject, ...project };
-        setProject(mockProject);
-        return mockProject;
-      };
       return (
         <TabbedModelForm
           {...args}
@@ -230,10 +90,6 @@ const meta: Meta<typeof TabbedModelForm> = {
       );
     },
   ],
-  beforeEach: () => {
-    mockModel = { ...model };
-    mockProject = { ...project };
-  },
 };
 
 export default meta;
